Rename planckunit submodule folder to planck-unit
