+++ .gitmodules
@@ -1,5 +1,5 @@
-[submodule "src/tests/planckunit"]
-	path = src/tests/planckunit
+[submodule "src/tests/planck-unit"]
+	path = src/tests/planck-unit
 	url = https://github.com/iondbproject/planck-unit
 	branch = iondb-sub
 [submodule "dotfiles"]
