+++ src/tests/behaviour/dictionary/behaviour_dictionary.h
@@ -34,7 +34,7 @@ extern "C" {
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "../../planckunit/src/planck_unit.h"
+#include "../../planck-unit/src/planck_unit.h"
 #include "../../../dictionary/dictionary_types.h"
 #include "../../../dictionary/dictionary.h"
 #include "../../../dictionary/ion_master_table.h"
