+++ src/tests/behaviour/dictionary/flat_file/test_behaviour_flat_file.c
@@ -21,7 +21,7 @@
 */
 /******************************************************************************/
 
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../behaviour_dictionary.h"
 #include "../../../../dictionary/flat_file/flat_file_dictionary_handler.h"
 #include "test_behaviour_flat_file.h"
