+++ src/tests/behaviour/dictionary/open_address_file_hash/test_behaviour_open_address_file_hash.c
@@ -21,7 +21,7 @@
 */
 /******************************************************************************/
 
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../behaviour_dictionary.h"
 #include "../../../../dictionary/open_address_file_hash/open_address_file_hash_dictionary_handler.h"
 #include "test_behaviour_open_address_file_hash.h"
