+++ src/tests/behaviour/dictionary/skip_list/test_behaviour_skip_list.c
@@ -21,7 +21,7 @@
 */
 /******************************************************************************/
 
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../behaviour_dictionary.h"
 #include "../../../../dictionary/skip_list/skip_list_handler.h"
 #include "test_behaviour_skip_list.h"
