+++ src/tests/unit/dictionary/flat_file/test_flat_file.h
@@ -31,7 +31,7 @@ extern "C" {
 #include <stdio.h>
 #include <string.h>
 #include <limits.h>
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../../../../dictionary/flat_file/flat_file.h"
 
 void
