+++ src/tests/unit/dictionary/flat_file/test_flat_file_dictionary_handler.h
@@ -30,7 +30,7 @@ extern "C" {
 
 #include <stdio.h>
 #include <string.h>
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../../../../dictionary/flat_file/flat_file_dictionary_handler.h"
 
 void
