+++ src/tests/unit/dictionary/open_address_hash/test_open_address_hash_dictionary_handler.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 #include <string.h>
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../../../../dictionary/dictionary_types.h"
 #include "./../../../../dictionary/dictionary.h"
 #include "../../../../dictionary/open_address_hash/open_address_hash.h"
