+++ src/tests/unit/dictionary/skip_list/skip_list_tests.h
@@ -11,7 +11,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "../../../planckunit/src/planck_unit.h"
+#include "../../../planck-unit/src/planck_unit.h"
 #include "../../../../dictionary/skip_list/skip_list_types.h"
 #include "../../../../dictionary/skip_list/skip_list.h"
 #include "../../../../dictionary/skip_list/skip_list_handler.h"
